stage 3 completed

+++ Recipe Management System/task/src/recipes/business_layer/dto/AddResponseDTO.java
@@ -1,4 +1,4 @@
 package recipes.business_layer.dto;
 
-public record AddResponseDTO(int id) {
+public record AddResponseDTO(long id) {
 }
